requirements.txt
django==1.10.3
django-admin-sortable2==0.6.5
django-ckeditor==5.1.1
gunicorn==19.6.0
pillow==3.4.2
psycopg2==2.6.2
redis==2.10.5
sorl-thumbnail==12.3
